Adding Armor Proficiency Parsing
App now parses armor proficiencies from classes and determines if the equipped armor is one the character is proficient in. If it is not, an exclamation point is added next to AC in the header and, when clicked, shows an alert explaining that the character is not proficient in the equipped armor and they, therefore, have disadvantage on Strength and Dexterity checks. I'm also working on extracting the API calls and logic from the context, so there are a few extra files in this commit.