Added ability to drop items (unfinished)
Added a button to each item to drop or take. Dropping the item reduces the carried weight; picking the item up increases the carried weight. Additionally, dropping an item unequips it. If the item is armor, it adjusts the AC accordingly. Currently, there is more than 1 of a single item, it drops all of them.